Move interrupt handling from interface to the shell
