+++ src/async_kernel/shell/ipshell.py
@@ -5,9 +5,12 @@ import json
 import pathlib
 import shlex
 import sys
+import threading
 import time
+import weakref
 from collections.abc import Callable
 from contextlib import contextmanager
+from sqlite3 import OperationalError
 from typing import TYPE_CHECKING, Any, Literal, Never, Self, TextIO, cast
 
 import anyio
@@ -16,9 +19,14 @@ from aiologic.lowlevel import async_checkpoint
 from anyio.streams.text import TextReceiveStream
 from IPython.core.completer import provisionalcompleter, rectify_completions
 from IPython.core.displayhook import DisplayHook
+from IPython.core.displaypub import DisplayPublisher
+from IPython.core.extensions import ExtensionManager
+from IPython.core.formatters import DisplayFormatter
+from IPython.core.history import HistoryManager, HistorySavingThread
 from IPython.core.interactiveshell import ExecutionResult, InteractiveShell
 from IPython.core.interactiveshell import _modified_open as _modified_open_  # pyright: ignore[reportPrivateUsage]
 from IPython.core.magic import Magics, line_cell_magic, line_magic, magics_class, no_var_expand
+from IPython.core.prefilter import PrefilterManager
 from IPython.display import display
 from IPython.utils.ipstruct import Struct
 from IPython.utils.tokenutil import token_at_cursor
@@ -30,14 +38,6 @@ import async_kernel
 from async_kernel import utils
 from async_kernel.caller import Caller
 from async_kernel.common import Fixed, KernelInterrupt
-from async_kernel.compat.ipython import (
-    IPDisplayFormatter,
-    IPDisplayHook,
-    IPDisplayPublisher,
-    IPExtensionManager,
-    IPHistoryManager,
-    IPPrefilterManager,
-)
 from async_kernel.compiler import XCachingCompiler
 from async_kernel.event_loop.run import get_runtime_matplotlib_guis
 from async_kernel.interface.base import BaseInterface, HasInterface
@@ -49,8 +49,11 @@ if TYPE_CHECKING:
     from types import CodeType
 
     from anyio.abc import ByteReceiveStream
+    from IPython.core.interactiveshell import ExecutionResult
 
     from async_kernel.pending import Pending
+    from async_kernel.shell import IPShell
+    from async_kernel.typing import Message
 
 
 __all__ = ["IPShell"]
@@ -71,6 +74,210 @@ class NullContext:
 
     def __exit__(self, type, value, traceback) -> Literal[False]:
         return False
+
+
+class IPDisplayHook(HasInterface, DisplayHook):
+    """Set as sys.displayhook and called whenever the interpreter needs to display output."""
+
+    cache_size = traitlets.Int(1000, min=3).tag(config=True)
+    do_full_cache = traitlets.Int(0).tag(config=True)
+
+    _ = __ = ___ = ""
+
+    def __init__(self, shell: IPShell) -> None:
+        self._shell_ref = weakref.ref(shell)
+        super(DisplayHook, self).__init__()
+
+    @property
+    def exec_result(self) -> ExecutionResult | None:  # pyright: ignore[reportImplicitOverride]
+        return None  # pragma: no cover
+
+    @exec_result.setter
+    def exec_result(self, exec_result: ExecutionResult) -> None:
+        pass  # pragma: no cover
+
+    @property
+    @override
+    def shell(self) -> IPShell:
+        return self._shell_ref()  # pyright: ignore[reportReturnType]
+
+    @property
+    @override
+    def prompt_count(self) -> int:
+        return self.shell.execution_count
+
+    @override
+    def start_displayhook(self) -> None:
+        pass  # pragma: no cover
+
+    @override
+    def write_output_prompt(self) -> None:
+        pass  # pragma: no cover
+
+    @override
+    def write_format_data(self, format_dict, md_dict=None) -> None:
+        pass  # pragma: no cover
+
+    @override
+    def finish_displayhook(self) -> None:
+        pass  # pragma: no cover
+
+    @override
+    def quiet(self) -> bool:
+        return not utils.show_result_enabled()
+
+    @override
+    def __call__(self, result=None) -> None:
+        """
+        Publish execution results.
+
+        This is invoked every time the interpreter needs to print, and is
+        activated by setting the variable sys.displayhook to it.
+        """
+        if result is not None and not self.quiet():
+            format_dict, md_dict = self.shell.display_formatter.format(result)
+            self.update_user_ns(result)
+            if format_dict:
+                content = {}
+                content["execution_count"] = self.shell.execution_count
+                content["data"] = format_dict
+                content["metadata"] = md_dict
+                self.log_output(format_dict)
+                self.parent.iopub_send("execute_result", content=content)
+
+
+class IPDisplayPublisher(HasInterface, DisplayPublisher):
+    """A display publisher that publishes data using [iopub_send][async_kernel.interface.base.BaseInterface.iopub_send]."""
+
+    _hooks: Fixed[Self, list[Callable[[Message[Any]], Any]]] = Fixed(list)
+
+    @override
+    def publish(  # pyright: ignore[reportIncompatibleMethodOverride]
+        self,
+        data: dict[str, Any],
+        metadata: dict | None = None,
+        *,
+        transient: dict | None = None,
+        update: bool = False,
+        **kwargs,
+    ) -> None:
+        """
+        Publish a display-data message.
+
+        Args:
+            data: A mime-bundle dict, keyed by mime-type.
+            metadata: Metadata associated with the data.
+            transient: Transient data that may only be relevant during a live display, such as display_id.
+                Transient data should not be persisted to documents.
+            update: If True, send an update_display_data message instead of display_data.
+
+        [Reference](https://jupyter-client.readthedocs.io/en/stable/messaging.html#update-display-data)
+        """
+        content = {"data": data, "metadata": metadata or {}, "transient": transient or {}} | kwargs
+        msg_type = "update_display_data" if update else "display_data"
+        msg = self.parent.msg(msg_type, content=content, parent=utils.get_parent_message())
+        for hook in self._hooks:
+            try:
+                msg = hook(msg)
+            except Exception:
+                pass
+            if msg is None:
+                return
+        self.parent.iopub_send(msg)
+
+    @override
+    def clear_output(self, wait: bool = False) -> None:
+        """
+        Clear output associated with the current execution (cell).
+
+        Args:
+            wait: If True, the output will not be cleared immediately,
+                instead waiting for the next display before clearing.
+                This reduces bounce during repeated clear & display loops.
+        """
+        self.parent.iopub_send(msg_or_type="clear_output", content={"wait": wait}, ident=b"display_data")
+
+    def register_hook(self, hook: Callable[[Message[Any]], Any]) -> None:
+        """Register a hook for when publish is called.
+
+        The hook should return the message or None.
+        Only return `None` when the message should *not* be sent.
+        """
+        self._hooks.append(hook)
+
+    def unregister_hook(self, hook: Callable[[Message[Any]], Any]) -> None:
+        while hook in self._hooks:
+            self._hooks.remove(hook)
+
+
+class IPHistoryManager(HasInterface[BaseInterface["IPShell"]], HistoryManager):
+    @property
+    @override
+    def shell(self) -> IPShell:
+        return self._shell_ref()  # pyright: ignore[reportReturnType]
+
+    @override
+    def __init__(self, *, shell: IPShell) -> None:
+        """Create a new history manager associated with a shell instance."""
+        self._shell_ref = weakref.ref(shell)
+        hist_file = ":memory:" if not shell.is_mainshell or sys.platform == "emscripten" else ""
+
+        super(HistoryManager, self).__init__(hist_file=hist_file)
+
+        self.db_input_cache_lock = threading.Lock()
+        self.db_output_cache_lock = threading.Lock()
+
+        try:
+            self.new_session()
+        except OperationalError as e:
+            self.log.exception(
+                "Failed to create history session in %s. History will not be saved.", self.hist_file, exc_info=e
+            )
+            self.hist_file = ":memory:"
+
+        self.using_thread = False
+        if self.enabled and self.hist_file != ":memory:":
+            self.save_thread = HistorySavingThread(self)
+            utils.mark_thread_pydev_do_not_trace(self.save_thread)
+            try:
+                self.save_thread.start()
+            except RuntimeError as e:
+                self.log.exception(
+                    "Failed to start history saving thread. History will not be saved.",
+                    exc_info=e,
+                )
+                self.hist_file = ":memory:"
+            else:
+                self.using_thread = True
+        else:
+            self.save_thread = None
+            if shell is not shell.kernel.main_shell:
+                self.output_hist.update(shell.kernel.main_shell.history_manager.output_hist)
+
+        self._instances.add(self)
+
+    def stop(self) -> None:
+        self.end_session()
+        if thread := self.save_thread:
+            thread.stop()
+            thread.join()
+            self.save_thread = None
+        self._instances.discard(self)
+
+
+class IPDisplayFormatter(HasInterface, DisplayFormatter):
+    pass
+
+
+class IPPrefilterManager(HasInterface, PrefilterManager):
+    pass
+
+
+class IPExtensionManager(HasInterface, ExtensionManager):
+    shell: IPShell
+
+    def __init__(self, *, shell: IPShell) -> None:
+        super().__init__(shell=shell)
 
 
 class IPShell(BaseShell, InteractiveShell):  # pyright: ignore[reportUnsafeMultipleInheritance, reportIncompatibleVariableOverride, reportIncompatibleMethodOverride]
@@ -263,7 +470,7 @@ class IPShell(BaseShell, InteractiveShell):  # pyright: ignore[reportUnsafeMulti
             self.parent.stop()
 
     def ask_exit(self) -> None:
-        if self.parent.raw_input("Are you sure you want to stop the kernel?\ny/[n]\n") == "y":
+        if self.kernel.raw_input("Are you sure you want to stop the kernel?\ny/[n]\n") == "y":
             self.exit_now = True
 
     @override
@@ -301,7 +508,7 @@ class IPShell(BaseShell, InteractiveShell):  # pyright: ignore[reportUnsafeMulti
         if msg := utils.get_parent_message():
             if "silent" in msg["content"]:
                 show = not msg["content"]["silent"]
-            if show and (code := msg["content"].get("code")) and DisplayHook.semicolon_at_end_of_expression(code):
+            if show and (code := msg["content"].get("code")) and IPDisplayHook.semicolon_at_end_of_expression(code):
                 show = False
         return utils.show_result(show)
 
@@ -444,7 +651,7 @@ class IPShell(BaseShell, InteractiveShell):  # pyright: ignore[reportUnsafeMulti
 
                 result = None
                 try:
-                    self.parent.interrupts.add(cancel)
+                    self._interrupts.add(cancel)
                     if stop_on_error:
                         self._stop_on_error_pool.add(cancel)
                     with anyio.fail_after(delay=timeout or None):
@@ -468,7 +675,7 @@ class IPShell(BaseShell, InteractiveShell):  # pyright: ignore[reportUnsafeMulti
                         err = RuntimeError(msg)
                 finally:
                     self._stop_on_error_pool.discard(cancel)
-                    self.parent.interrupts.discard(cancel)
+                    self._interrupts.discard(cancel)
                     self.events.trigger("post_execute")
                     if not silent:
                         self.events.trigger("post_run_cell", result)
